+++ frontend/01_connect_and_explore.py
@@ -21,7 +21,7 @@ import polars as pl
 import streamlit as st
 from streamlit.runtime.uploaded_file_manager import UploadedFile
 
-sys.path.append("..")
+sys.path.append(os.path.dirname(os.path.realpath(__file__)))
 from app_settings import (
     apply_custom_css,
     display_page_logo,